[UPDATE] Solução dos Problemas 9 e 10
Solução dos Problemas 9 e 10.
Mudança do tipo do retorno do método Solve de Int32 para Int64

+++ src/ProjectEuler/IEulerProblem.cs
@@ -3,6 +3,6 @@ namespace ProjectEuler
 {
     public interface IEulerProblem
     {
-        int Solve();
+        long Solve();
     }
 }
